Add sticky search navbar
